requirements.txt
# Core
streamlit>=1.32.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Gmail API
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.126.0

# Legacy (kept for compatibility with existing scripts)
gspread>=5.0.0
gspread-dataframe>=4.0.0
watchdog>=3.0.0
